requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
pydantic==2.7.4
pydantic-settings==2.10.1
